requirements.txt
discord.py
astncio